requirements.txt
anyio==4.10.0
APScheduler==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
pymongo==4.15.0
python-dotenv==1.1.1
python-telegram-bot==22.4
sniffio==1.3.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
